Remove the category pickers and tag pickers
