+++ app/assets/stylesheets/records.sass
@@ -53,10 +53,6 @@
     .charge-td
       width: 80px
 
-.picker-label
-  display: block
-  padding: 5px 10px
-
 .picker-button
   background-color: $rgba7-white
   border: solid 1px $green
