--- app/controllers/api/recently_used_controller.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class Api::RecentlyUsedController < Api::BaseController
-  def index
-    render json: current_user, include: '**'
-  end
-end
