+++ app/front/javascripts/NewRecordCard.js
@@ -16,7 +16,7 @@ class NewRecordCard extends React.Component {
     return (
       <div className='new-record-card-component'>
         <ErrorBoundary>
-          <NewRecordCardBody recentlyUsed={this.props.recently_used} records={this.props.records} />
+          <NewRecordCardBody records={this.props.records} />
         </ErrorBoundary>
       </div>
     )
@@ -25,7 +25,6 @@ class NewRecordCard extends React.Component {
 
 NewRecordCard.propTypes = {
   records: PropTypes.array.isRequired,
-  recently_used: PropTypes.object.isRequired,
   user_token: PropTypes.string.isRequired,
   last_request_at: PropTypes.number.isRequired
 }
