--- app/front/javascripts/mixins/requests/RecentlyUsedMixin.js
@@ -1,8 +1,0 @@
-import { axiosMixin } from './AxiosMixin'
-
-export const recentlyUsedAxios = {
-  get : (callback, errorCallback) => {
-    let url = '/api/recently_used'
-    axiosMixin.request('GET', callback, errorCallback, url)
-  }
-}
